feat: Create modules for main.js file. Fix naming for variables (UpperCase)
