+++ js/create-author.js
@@ -1,0 +1,9 @@
+import {createCounter} from './util.js';
+
+const getCounter = createCounter();
+
+const createAuthor = () => ({
+  avatar: `img/avatars/user${getCounter().padStart(2,0)}.png`
+});
+
+export {createAuthor};
